Add unique constraint to OrganisationPermission model
At the moment, there's nothing to stop an organisation having lots of
rows for the same permission in the database. While the frontend code
shouldn't be adding duplicate rows let's make sure that scenario doesn't
occur and give unexpected behaviour.

If creating the table from scratch this could have worked like the
`service_permissions` table, with the organisation_id and permission columns
forming a composite primary key. However, since the table exists it's
a lot of effort to change it to that format.

There is only one type of organisation permission available,
`can_ask_to_join_a_service`, and it's for a feature that has not
launched, so making this change now is safe - the
organisation_permission table only has a couple of rows of data.