Add new economy mail rates for April 2025
Currently we offer letter postage options for First Class, Second Class, Europe, and the Rest of the World. We are now adding economy postage rates for April 2025.
We recently updated the rates for our existing postage options (start date 31-03-2025 23:00), so we're using the same start date for the new economy rates.