Add renamed route for getting a service join request
A new route with the service id in the URL had been added, and the old
route can be removed once the admin app has been changed to not use it.

This also adds an additional check to the DAO function that the join
request is for the specified service. If no result is found, the error
handler will deal with ensuring a `404` status is returned to the admin
app.