Add new route for updating a service join request
This adds a new route which does the same as the existing one, but which
has the service id in the URL. The case where a join request does not
exist is now handled automatically by the error handlers, so doesn't
need to be checked in the view function.

The existing tests have been updated for the new route, since the old
route will be removed.