Merge the endpoints to create a service join request
This moves the functionality of the
`/service/<service_id>/invite/request-for/<user_to_invite_id>` endpoint
into the `/service/<uuid:service_id>/service-join-request`, since this
is the version we want to keep. Some minor changes are needed, since the
data each expected was different. It also copies the tests for the
deprecated endpoint so that they work with the new changes.