Part 1 of refactor to validate and get phone numer data in validators
instead of in persist_notifications

This is to:
- only validate once, not twice
- keep persist_notifications light and keep validation out of it
- so that we can check international sms rate limits in validators
for all routes of sms, including jobs.

Since it's a big refactor, I have split it into partial commits.

In part 1 (out of 6):
Update validate_and_format_recipient() function:
- split getting phone number info into a private sub-function
of validate_and_return_extended_phone_number_info().
- return extended phone info as a dict - this is all the
phone number info that persist_notification() needs.

I know this is in some ways not ideal, as it means we return
a string when validating email addresses and a dict
when validating phone numbers. At the same time I think the
benefit of not validating twice, and of moving the validation
logic out of  persist_notification() outweighs this.

I'd be open to refactoring it more at some point.