Part 2 of refactor to validate and get phone numer data in validators
instead of in persist_notifications

This is to:
- only validate once, not twice
- keep persist_notifications light and keep validation out of it
- so that we can check international sms rate limits in validators
for all routes of sms, including jobs.

Since it's a big refactor, I have split it into partial commits.

In part 2 (out of 6):
Update persist_notification() so it can accept recipient
argument either as a string or as a dict, and so for SMS
it doesn't do validation - instead it receives all SMS recipient
data it needs.