Part 3 of refactor to validate and get phone numer data in validators
instead of in persist_notifications

This is to:
- only validate once, not twice
- keep persist_notifications light and keep validation out of it
- so that we can check international sms rate limits in validators
for all routes of sms, including jobs.

Since it's a big refactor, I have split it into partial commits.

In part 3 (out of 6):
Update one-off text notification processing code to receive
the dict with recipient data from validators and pass it on to
persist_notification() function