Part 5 of refactor to validate and get phone numer data in validators
instead of in persist_notifications

This is to:
- only validate once, not twice
- keep persist_notifications light and keep validation out of it
- so that we can check international sms rate limits in validators
for all routes of sms, including jobs.

Since it's a big refactor, I have split it into partial commits.

In part 5 (out of 6):
Update the code processing sms notifications from jobs to validate
phone number and pass the dict with phone number data on to
persist_notification() function.
It also does what persist_notification did before - when validation
fails, it still sends the sms to be persisted, just with no billable
units and in validation_failed status.