Part 6 of refactor to validate and get phone numer data in validators
instead of in persist_notifications

This is to:
- only validate once, not twice
- keep persist_notifications light and keep validation out of it
- so that we can check international sms rate limits in validators
for all routes of sms, including jobs.

Since it's a big refactor, I have split it into partial commits.

In part 6 (out of 6):
Update the code that sends sms notifications from Notify service,
like when we send 2FA codes to useres. Now this code validates
the phone number and then sends recipient data dict to persist_notifications().
(earlier it would send the number to persist_notifications() and it would
get validated there).