Update from productionRunDate to transitionDate
Since February 2025, we have noticed a mismatch between the GDS ft_billing report and the DVLA billing report. After looking into it, we found that the issue is linked to a callback change we made at the end of 2024. At that time, we were told to use productionRunDate (Print Date), but the problem started when DVLA changed their report or began using transitionDate instead.

The change we are now making updates the despatched_on value from productionRunDate to transitionDate.

When DVLA sends a callback, details like notification_id, despatched_on, and cost_threshold are stored in the notifications_letter_despatch table. A nightly task then runs to calculate billing and stores it in ft_billing.

Each time we get a callback for a letter, we receive a despatched_on date (previously productionRunDate). If we switch to using transitionDate, it would not affect billing, because the despatched_on date is stored with the notification, and the billing cron job reprocesses new back dated data from the last 10 days.