Check economy mail service permission when sending via API
This adds a check that a service trying to send an economy letter has
the `economy_letter_sending` service permission. If not, we stop the
letter notification from being created.

We check most service permissions at the initial stage of making the API
request, but we don't know the postage until a bit later in the process
so can only do this check when creating the notification.