Treat crown dependency numbers as domestic when checking limits
When checking and adding to  daily international sms limit.

We introduced international sms daily limit, and it includes all
international sms, including crown dependency numbers.
I included them on purpose, because some of the AIT attacks have been
coming from an MNO on Jersey. What I did not foresee, is that it would
take us out of sync with the permission check that checks if a service
can send to international sms at all. This check treats crown dependency
numbers as domestic. Even though rate multiplier for them is x2.
So now we have a user from a service with international_sms turned off,
who cannot send to what looks to her as a bunch of domestic numbers,
because of the new check.
For now we are working to harmonise the new check to the old check - that is
treat crown dependency numbers as domestic.
But that means our services will still be vulnerable to the AIT attacks
from these locations.
So I am in two minds how we go forward - do we just keep treating
crown dependencies as domestic to avoid confusing our users
(since the prefix and number format is the same as a domestic number),
or since they are more expensive and related to historic AIT attacks,
we should treat them as international and let our users know?

We will discuss these questions as a team, but for now lets
harmonise the checks, so users can send the messages they
expect to be able to send.