Stop incrementing overall message limit for services
We stopped checking the overall message limit in this PR:
https://github.com/alphagov/notifications-api/pull/3706/files

We dropped the database column entirely in this migration:
https://github.com/alphagov/notifications-api/pull/3712/files

This was part of the work to move to a per-channel message limit.

However we are still keeping track of how many messages each service
has sent each day in Redis, even though we never look at this value.

This is an unneccessary extra call to Redis each time a message is sent,
and confuses people reading the code.