Fix report request size
Without seek(0), the CSV ends up with a lot of extra empty space at the end. This makes the file much larger than it should be and slows down report generation. Adding self.csv_buffer.seek(0) moves the cursor to the start of the buffer before reading from it. This makes sure the buffer is fully cleared so the next batch of data starts clean.