Fix batch time for report request
For large datasets, using page and skip slows down the query because it needs to count and skip many rows. Cursor-based pagination is faster since it jumps directly to the next item using a cursor. It works by taking the notification ID from the last batch and passing it as older_than to get the next set. This avoids scanning unnecessary data and also used by get notifications API.