Remove unused service permissions
All of these were for feature flags which have since been rolled out to
all services so are no longer being checked. The admin part of the work
was done in https://github.com/alphagov/notifications-admin/pull/5462

This change was initially partly rolled out in
https://github.com/alphagov/notifications-api/pull/4464, which caused
issues when deployed, so needed to be reverted. In that case, we still
had services with the old service permissions in the database, but had
removed references to them in the code. This caused a schema validation
error when updating a service which had one of the old permissions.

The database change and code change are now grouped together. During
deployment, the database migration will run first, removing the old
permissions from services. The permissions will then be removed from the
code, which should prevent the same issue.