Stop giving new services UPLOAD_LETTER permission
We don't check this service permission, it existed as a feature flag
when we were developing the upload letter feature, but now all services
can upload letters
(https://github.com/alphagov/notifications-api/pull/2735).

As a first step towards removing it from the
database we need to ensure we stop giving it to every new service that
gets created.