Add job_id index to notification_history
This is a temporary change to allow us to delete the old emergency
alerts data. The `job_id` foreign key constraint on the
`notification_history` table means that when attempting to delete the
jobs, Postgres also needs to check the `notification_history` table
which is taking too long without an index.

It was not possible to send a broadcast as a job, but there are still
broadcast services in the database which have rows in the `jobs` table
because in some cases, we converted an existing "normal" service to a
broadcast service.