Add service_id index to notification_history
This is a temporary change to allow us to delete the old emergency alerts data.
The service_id foreign key constraint on the notification_history table means
that when attempting to delete the services, Postgres also needs to check the
notification_history table which is taking too long without an index.