Remove emergency alerts data
This is a data migration (so no downgrade) that deletes broadcast
services and all associated data - this includes obvious things like
broadcast events and templates, but also everything connected to those
services.

Although some of these theoretically shouldn't exist for broadcast
services (eg returned_letters), we attempt to delete them out of an
abundance of caution.

This work is based off an older PR, which was tested against a copy of the
staging database. That PR was reverted because it was taking too long to
delete from the `notification_history` table due to the foreign key
constraints - since then, indexes have been added to
notification_history to make deleting from the table faster.

Differences between this PR and the original are:
* the `service_inbound_api` table no longer exists, so has been removed
  from the migration
* this adds the `api_keys_history`, `service_callback_api_history`,
  `ft_billing` and `ft_notification_status` tables to the list we delete
  from
* this deletes from the `notifications` and `notification_history` tables
  first - these have various foreign key constraints on other tables
  which would cause issues if the data from those other tables is deleted
  first.