Update SQLAlchemy mapping to use new registry API
The old mapper() function which was deprecated and then removed in SQLAlchemy 2.0. The sqlalchemy.orm.mapper() standalone function moves behind the scenes to be invoked by higher level APIs. The new version of this function is the method registry.map_imperatively() taken from a registry object. See https://docs.sqlalchemy.org/en/20/changelog/migration_20.html\#the-original-mapper-function-now-a-core-element-of-declarative-renamed for more details.