Update QueryRecorder to use db.get_debug_queries for Sqlalchemy 2.0
The get_debug_queries function was removed from the top-level flask_sqlalchemy module in a recent version of the library and is now a method on the SQLAlchemy 2 instance (db). This change updates the QueryRecorder test utility to accept the db instance and call db.get_debug_queries() instead of the old, removed function.