Change timing of daily check for letters pending virus scan
This was taking place at 2000 UTC, meaning that if any letters were
found to still be pending virus check it was too late for them to make
the evening's print run (which has a cut of time of 1730 local time).

Moving the check earlier means that any letters this task picks up
should be able to make the print run for the day. I've also extended the
number of days we check by one, since we did see once case where this
number of attempts wasn't long enough.