Use `TemplateHistorySchema` to serialise template versions
 ## Admin app

Gets a serialised template version from here:
https://github.com/alphagov/notifications-api/blob/12b0aebfca0d2a1506cb208c12a58022b52c9c40/app/template/rest.py#L206-L208

Caches this in Redis here:
https://github.com/alphagov/notifications-admin/blob/cdd2286251446a04d2ae885cfc0da79acc39cb9e/app/notify_client/service_api_client.py#L256

 ## API

Gets serialised template version from here:
https://github.com/alphagov/notifications-api/blob/12b0aebfca0d2a1506cb208c12a58022b52c9c40/app/serialised_models.py#L69

Caches this in Redis here (against the same key used by the
admin app):
https://github.com/alphagov/notifications-api/blob/12b0aebfca0d2a1506cb208c12a58022b52c9c40/app/serialised_models.py#L58

***

The API is using:
- `TemplateHistorySchema` to serialise a template version for the admin app
- `TemplateSchema` to serialise a template version for its own use

This means that, depending on who gets there first, data will be cached
in different formats against the same key.

This commit makes sure that:
- the current version of a template is always serialised using `TemplateSchema`
- previous versions of a template are always serialised using `TemplateHistorySchema`

This should make no difference to the data the API is using, because
`SerialisedTemplate` restricts which attributes can be accessed, and
none of these attributes are the ones which differ between the two
schemas.