Remove unused `check_reply_to` function
This was added in https://github.com/alphagov/notifications-api/pull/1500/files
for the internal API, as part of the work to add contact blocks for
letters.

The original intention was that a default reply-to could be set on any
type of template. We never went through with this and to this day only
letters can have a per-template reply-to address (the contact block).

To clean things up we stopped using the generic function in
https://github.com/alphagov/notifications-api/pull/4037/files

We still use the letter-specific function in 2 places on the internal
API:
- https://github.com/alphagov/notifications-api/blob/6857eeac6fdf8f54c20f9016c30aa5a49f1b9d7d/app/template/rest.py#L106
- https://github.com/alphagov/notifications-api/blob/6857eeac6fdf8f54c20f9016c30aa5a49f1b9d7d/app/template/rest.py#L131

We only call the underlying, per-channel functions directly when sending
a notification:
https://github.com/alphagov/notifications-api/blob/6857eeac6fdf8f54c20f9016c30aa5a49f1b9d7d/app/v2/notifications/post_notifications.py#L385-L407

So we can safely remove this generic function.

I’ve updated the tests so the per-channel functions still have the same
coverage.