Promptly update the status for jobs that completed processing
Every minute, run a task that updates the status for jobs
that finished processing all rows to JOB_STATUS_FINISHED_ALL_NOTIFICATIONS_CREATED.

This is so cancelling letter jobs works as expected.

Before this PR, users would have to wait around 30 minutes after
sending a job before being able to cancel letters. This is a
confusing experience, and it has led to a user thinking they cancelled
their letters when they have not.