Convert debug type loglines into info type loglines
Debug type loglines do not result in production logs.
Devs may accidentally use this type of logline for production code,
which is what I think happened to these log lines. Then, if there
is an issue with the code where the logline is, and the log could
have helped, we are left in the dark isntead.

This happened recently. When the
code automating unsubscribes for Notify users has been merged, it
included a helpful logline letting us know the unsubscribe has
been processed. Unfortunately the code had a bug - the db session
was not committed, and the unsubscribe info was not persisted
in the database. Unluckily, we also sent a big round of research
comms in the same week. By the time we found the bug, there would
have been multiple unsubscribes. We tried to find how many, with
the helpful log that the code was shipped with. Unfortunately,
the logline was added as a debug type, meaning it did not result
in any production logs. That meant we wouldn't even know how
many unsubscribes we lost.
Luckily, this is just a near miss - because by accident, that
round of comms was not sent from the right Notify service, and so
the unsubscribes went through the manual flow instead.

Still, a near miss like this is an important lesson - we need to
remember to add logs to our code, and these logs should be of type
info, warning or error - never debug for code going to production
- unless you have a good reason for it and explain it in a comment
directly above that logline - so other devs don't copy your
pattern by mistake.

It is possible that some of the loglines in this commit are
redundant - for example we might have similar logs about notification
being created elsewhere in the code, and we might not need both.
Since we send millions of notifications every day, each such logline
will result in millions of extra logs. In that case we should either
totally remove these loglines, or move them back to debug type, with
an appropriate comment.