add deep-archive-notification-history-up-to-limit celery task
this celery task will look for the oldest notifications in
NotificationHistory and, in hour-sized chunks, "archive" them
to ORC files in S3. if NOTIFICATION_DEEP_HISTORY_DELETE_ARCHIVED
is enabled, it will also then attempt to delete the rows it
has archived from the table. if the deletion is successful it
will add a tag to the S3 object marking it as
contents_deleted=true.

it will do this for a maximum
NOTIFICATION_DEEP_HISTORY_MAX_HOURS_ARCHIVED_IN_RUN hours in
a single run, and will not proceed past NotificationHistory
rows more recent than NOTIFICATION_DEEP_HISTORY_MIN_AGE_DAYS.

the intent of allowing NOTIFICATION_DEEP_HISTORY_DELETE_ARCHIVED
to be disabled is to enable initial performance tests on
a production system to be run without making them "final" - i.e.
so that their output is still "disposable" if we decide changes
need to be made to run this regularly in production.

with deletion disabled it will of course not be a 100% faithful
representation of the performance we should expect, because
the deletion itself will take up a significant amount of
database resources and the exclusive locks that are taken
by the deletion command may cause blocking or even deadlocking
with other database clients. the theory is that this is
unlikely to be a problem because very few parts of the app
access the very old regions of NotificationHistory. fewer still
attempt to update these rows, which is what would cause
deadlocks.

still, the archiving process is designed to be robust to
encountering deadlocks - if the deletion is unsuccessful, the
S3 object will not have its tagging updated to
contents_deleted=true which should allow a lifecycle rule to
put a delete marker on it after a day (which should cause
athena to skip over it, but allow us to recover that file
if e.g. only the final-tagging operation failed after having
successfully performed the deletion). the celery task would
simply attempt to archive the same hour the next time it is
run.

care is taken to ensure that we don't delete rows that we
haven't included in the archive, and that the version of the
row that we archive is the "final" version of the row that
the database saw.

because we also always look for the oldest notifications in
NotificationHistory when selecting the next hour to archive
(rather than iterating through hours sequentially) we should
also be robust to notifications unexpectedly "leapfrogging"
the archival process if something caused one to be added from
far in the past. this "leapfrogging" notification would
simply cause another archival of the hour in question and
the resulting S3 object would be placed next to the original
archival of that hour. athena should be able to handle
multiple files for the same created_at_hour without trouble.

another note here is that we use the "hive-style" partitioning
scheme for the paths in S3. interestingly created_at_date_hour
should be considered separate from created_at - athena won't
assume any relationship between the two and when using athena
created_at_date_hour will need an explicit clauses in order
for the partitioning to have the desired effect. this is
because partitioned variables are supposed to perform best when
they are considered as pure strings.

while the ultimate intention is to hook this up to a celery
beat schedule, this has not been done in this PR - we will
probably want to manually trigger some initial test runs
in production before we're happy to leave it to run
"unattended"