Add resanitise letter attachment task
Add celery task to re-sanitise template letter attachments.

This task calls the  template preview task which recreates the PDF for a letter attachment which is already sanitised and in the letters-attachment bucket. The new file that is generated will then overwrite the existing letter in the letter-attachments bucket.

`service_id` is for a PDF letter attachment/template. This will be used to store the final sanitised PDF in the letter-attachment S3 bucket, where PDFs are organised into service-specific folders named after the service_id.
`attachment_id` is for a PDF letter attachment which was uploaded for a template. Letter attachment PDF files are stored using the attachment_id as the file name.
`attachment_name` is for a PDF letter attachment. In the letter-backup S3 bucket, PDF files are stored using only the attachment_id and no metadata. To make this easier, we can fetch the details from the database and pass them to Template Preview, so that the original file name is stored as metadata in the letter-attachment S3 bucket.