Fix `SQLALCHEMY_DATABASE_URI` in Docker script
The value of `SQLALCHEMY_DATABASE_URI` was changed to include the driver
and port explicitly in 5795537 and this also changed the host in the URI.
However because this script is used for running in Docker, the database host
needs to stay as `host.docker.internal` (the Docker host) and can't be
changed to `localhost`, which would refer to the container.

This change fixes things if running the app using Docker and without
notifications-local.