Don’t cast `replenish_rate_per_sec` to `int`
If a service has a very low rate limit (< 60 per minute) this will cause
their replenish rate to be rounded down to 0.

So once they have consumed the initial 1,000 tokens in their bucket
they never get any more.

The underlying Lua script is designed to handle fractional replenish
rates, so we can give it a `float`.