Reduce bucket max for services with low rate limits
We have services with very low rate limits to protect themselves from
automated attacks.

Giving them a bucket max of 1,000 means that they will accept at least
1,000 requests before starting to rate limit.

For a service with a limit of – for example – 10 per minute, this is a
big increase in potential requests.

This commit makes the bucket max proportional to the rate limit, for
services with a rate limit lower than the default. So they still get
some buffer, but not so much it could incur them unexpected expense.