add custom session to allow database read-replica routing
The RoutingSession is a custom database session to send certain database reads from read-replica.
Imported the standard Session from Flask-SQLAlchemy. This is the blueprint for the normal database connection.
It works by adding a using_bind(replica) function that we can call to tag the session. Then, it overrides the get_bind method, which checks for this tag before running a query. If it sees the replica tag, it sends that one query to the replica database and immediately removes the tag, so all other queries automatically go to the main database.