Remove bespoke provider cache
We are now caching the whole list of providers in memory.

This means we can stop caching the outcome of the choice.

This is better because it will give us a smoother distribution between
providers, rather each instance locking on to one for 10 seconds, then
lurching entirely to another one.