created post endpoint for creating template_email_file records
to be used by called by admin when one or more files are added to a template. Admin should be in control of setting the template id when we create the file record vs updates that should alwys bump the template version so as not to break templates_history and template_email_files history. This endpoint should however be restricted to email templates only