GET endpoints: get all by template_id, get latest by file_id, get by template_id and template_version
Three GET endpoints. Admin will need to get a list of all the files of a given version to do validation etc (e.g. if a user tries to delete a file placeholder, we need to display a warning etc). We need to get templates by id for the preview view. We may need to get historical endpoints too.