UPDATE endpoint - update files by id
uses `version_class` decorator to update both the template version and the file version when a change to the file is made. This is needed because it should be possible to uniquely map a template version to the files associated with it. This prevents jobs from sending the wrong files