add service permission for sending files by email via the notify web ui
in order to stage the deployment, run private betas etc. etc. with the send file by email via ui feature we need a new service permission that lets us add this feature under a feature flag. Long term we will probably just use the standard permission for sending files by email and let users use both the api and the UI for this.