Add column `confirmed_email_sender_name` to services
We want to record whether a service has confirmed that they're happy
with the email sender name the service will use. We can tell if they've
picked a custom email sender name (which they will be happy with since
they chose it specially) but not whether they have confirmed that they
are happy to use the service name as a sender.