Update template response to include a list of template_email_file names
This adds an additional key/value pair to templates GET endpoints that will:
- always return an empty list for sms and letter templates
- return an empty list for an email template with no files attached
- return a list of one or more files for an email template with files attached

Returning a list of filenames means we can reuse all of our caching logic for templates in both api (e.g. in serialised models) and admin, storing this within the cache keys for templates. The list of files is added as an attribute at the point of serialisation, it is not stored as an attribute/relation in the SQLAlchemy model.

We can't store this attribute in the templates table, and  it can't be found by using a JOIN: Templates and TemplateEmailFiles are loosely coupled whereby one or many TemplateEmailFiles row can be associated with 1 or many versions of a template. Relational databases don't properly model this, logic in our dao functions is instead used for querying and aggregating data. This isn't ideal, but we couldn't think of a properly relational way to do this that didn't otherwise break normalisation etc.