db connections: don't decrement TOTAL_DB_CONNECTIONS if we didn't increment it
our connection can fail before it manages to increment
TOTAL_DB_CONNECTIONS - add a new connection record variable
to track this and conditionally skip decrementing.

this problem manifested itself in some connections calling the
close() event handler while not having inet_server_addr set
in the connection record, resulting in a KeyError.

it appears we're most likely to get these early failures from
replica connections as there are a myriad of things that can
upset postgres on a replica and cause a "conflict" of some sort
to be raised.