Backfill some data in the confirmed_email_sender_name column
For new services, `confirmed_email_sender_name` will default to `False`
so will not be empty. For existing services,
`confirmed_email_sender_name` is null since we don't know if they have
seen the email sender name and confirmed they are happy with it.
However, the one case where we know they definitely have checked the
sender name is if they are using a custom sender name. This migration
backfills the columns in the services and services_history table to set
`confirmed_email_sender_name` to `True` in that scenario.