update process_notifications to optionally render linktext as a markdown link if link text has been set for the file
we provide users with the option of setting link text when sending files by email via the ui. If link text has been set, process notifications should wrap the secure link returned from the document-download-api in markdown so that a clickable link with the link text is rendered.

e.g. `https://example.com` --> `[click this link](https://example.com]` if the link text has been set to `click this link`