Archive email files when updating template
I tried doing this in a single session, but run into issues
with session flushing before updating history, resulting in history
tables not being updated. Hence, I kept it in one endpoint,
but two db sessions - one to update template, and one to archive
all the email files that need archiving.

I also don't update template version in the session where the
email files are updated - as the version already gets updated
when the template is updated. That saves us vacuous rows in the
template_history table.